Changed SG_TEST_COMM enum from READ to WRITE
Before it was right from a testing perspective: with
SG_TEST_COMM_READ_SPI, we are reading from the named spi pipe.
Now, it is right from an implementation perspective: with
SG_TEST_COMM_WRITE_SPI, we are writing to the serial line (and for
testing, we'll be reading that fifo). They both mean the same thing,
the WRITE just makes more sense