Added CK_NOFORK (commented out) for debugging
This is a check.h option; normall check forks tests, so that a segfault
doesn't bring down the entire test suite. This is not very good for
debugging, though, so there is an option to keep check.h from forking
off. I added it (commented out) for when I need to debug segfaults