Redifined heartbeat sleep inteval in seconds
the current sleep interval is 4 minutes. I was previously using
sleep_nano to sleep, but I was overflowing the variable that stored the
interval. So I changed the sleep interval to be in terms of seconds,
and I used sleep() instead. Later, if I have the time, I'll rework to
use sleep_nano again