Major progress: Built (terrible) router-side API, finished Eyeball, Peck, Sing, nd Mock (woohoo!), added ability to toggle automatic-mating.
Minor progress:

Broke out API functions to a separate page so that they may be
called from other pages (API.php has checks to ensure only the router may contact it).

Wrote a little script to process an entire series of payloads to simulate a live sensor.

Fixed a bunch of errors in my naming convention to conform to the intrarouter API.

Updated generate_stub.pl to generate a Mock.