Router allows no authentication
Note that right now this just means that the system won't complain
until the packet is prepped to be sent. Then it complains and doesn't
send. This functionality may not be implemented, so for now auth=none
has undefined behavior. Later it'll cause the sansgridrouter to act
exactly like a switch